+++ tools/Beef.CodeGen.Core/CHANGELOG.md
@@ -5,6 +5,8 @@ Represents the **NuGet** versions.
 ## v2.1.6
 - *Enhancement:* Added `WithHints` to stored procedure configuration to output `WITH(value)` table [hint](https://docs.microsoft.com/en-us/sql/t-sql/queries/hints-transact-sql-table).
 - *Enhancement:* Added support to generate data logic access using *Cosmos* DB. This follows the same pattern as *Database*, *OData* and *EntityFramework*. 
+- *Fixed:* `IEntityData` code-gen did not correctly output the value type.
+- *Enhancement:* Added code-gen support for snake_case and kebab-case.
 
 ## v2.1.5
 - *Fixed:* `InvokerBase` was non functioning as a generic class; reimplemented. Invoker invocation code generation updated accordingly.